chore: move wds example into its own folder
