+++ .gitignore
@@ -20,4 +20,4 @@ yarn-error.log*
 
 # lockfiles
 package-lock.json
-yarn.lock
+./yarn.lock
