+++ examples/wds-kitchen-sink/babel.config.js
