+++ examples/wds-kitchen-sink/package.json
@@ -20,7 +20,6 @@
     "cross-env": "^6.0.3",
     "html-webpack-plugin": "^3.2.0",
     "react-refresh": "^0.7.0",
-    "react-refresh-webpack-plugin": "https://github.com/pmmmwh/react-refresh-webpack-plugin#master",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.9",
     "webpack-dev-server": "^3.9.0"
