+++ examples/wds-kitchen-sink/src/ClassDefault.jsx
