+++ examples/wds-kitchen-sink/src/ClassNamed.jsx
