+++ examples/wds-kitchen-sink/src/FunctionNamed.jsx
