+++ examples/wds-kitchen-sink/src/LazyComponent.jsx
