+++ examples/wds-kitchen-sink/src/index.js
