+++ examples/wds-kitchen-sink/webpack.config.js
@@ -1,5 +1,5 @@
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ReactRefreshPlugin = require('../src');
+const ReactRefreshPlugin = require('../../src');
 
 module.exports = {
   entry: './src/index.js',
